backend/requirements.txt
flask==2.2.5
GitPython==3.1.32
neo4j-driver==5.11.0
qdrant-client==0.10.2
tree_sitter==0.20.0
transformers==4.30.0
torch==2.0.1
